Update schema handling
Revised load.py so that it passes through the schema defined in the
schema directory for a given version of HAND. It will aggregate into
Hydrotable by hydroid depending on if the column in the duckdb
Hydrotable schema has an array type or not.

Updated Readme.

Created a script analyze_hydrotable_columns.py that should be able to
give you an indication of if a column should become an array type or not