Fix/mem leak (#5)
* Update code to not leak memory

load.py has been updated to use duckdb for catchment gpkg processing and
for hydrotable aggregation. The query and schema have also been updated
in some places when necessary to work with minor differences in the
parquet files produced this way compared to the previous parquet files

* Remove example-hydrotable.csv

* optimize centroid calc

* Merge geometries

* Remove insertion order setting since closing connection now

* Reduce default batch size and document sizing guidelines

* Fix tolerance to be 100m

* Update gitignore

* Update load.py comments

* Add staging for hydrotables. Remove json columns.

* Add staging for catchments and modify raster table schema

* Replace WKT writing with WKB writing for catchment geometries

* Add more explanatory comments

* Remove broken batch insertion print statements and estimates

* Fix broken row insertion counts printout after table creation

* Add branch lookup table to improve performance

Create a branch lookup table that is used during hydrotable creation.
This table is indexed by branch so should be faster than using the
original catchment table.

* Aggregating before joining in hydrotables

* Flush staging hydrotable to avoid runing out of memory

* Fix parquet index geometry handling to deal with wkb format

Fixed the geometry loading in the test query script and in the
visualization script

* Fix new load.py new partition_tables_to_parquet

partition_tables_to_parquet now partitions files one H3 index at a time
to avoid running out of memory. Was previously not actually writing a
file so query wasn't running sucessfully

* Squash merge feature/simplify-hydrotable into fix/mem-leak. This gets rid of complex hydrotable handling in favor of just passing refs to the existing hydrotable csv's. This change reduced code complexity alot in this repo while requiring only minor changes to the autoeval coordinator.